+++ extensions/doctor/CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log
 
+# 1.1.6
+
+- feat: add showInformationMessage for codemod result
 
 # 1.1.5
 
 - fix: windows get empty files
+
 # 1.1.4
 
 - fix: codemod 100 score show result