+++ packages/material-utils/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 0.1.2
+
+- feat: add ts type(componentType)
+
 ## 0.1.1
 
 - feat: add languageType type to IMaterialScaffold